fix: address PR review feedback — TrialEvent topic, byte order, docstring, architecture diagram
- base.py: publish TrialEvent on TOPIC_EVENT (not TOPIC_TRIAL) so DataLoggerProcess captures all behavioral events on a single subscription
- messages.py: mark TOPIC_TRIAL as reserved with a comment
- data_logger_process.py: fix _write_haptic_sample to use explicit little-endian dtype ('<f8'), update docstring to say 'sample count' not 'sample rate'
- docs/architecture.md: remove duplicate DataLoggerProcess from Tier 2 (SyncProcess box); keep only in Tier 3
- test_task_session.py: filter TrialEvent by __msg_type__ instead of TOPIC_TRIAL (now arrives on TOPIC_EVENT)
- test_data_logger.py: use '<f8' (little-endian) dtype in frombuffer assertions

Agent-Logs-Url: https://github.com/raeedcho/hapticore/sessions/85d67d1c-a9fd-4533-9bac-515405fd3bbd

+++ tests/unit/test_data_logger.py
@@ -153,7 +153,7 @@ class TestWriteHapticSample:
         )
         DataLoggerProcess._write_haptic_sample(buf, msg)
         buf.seek(0)
-        row = np.frombuffer(buf.read(), dtype=np.float64)
+        row = np.frombuffer(buf.read(), dtype="<f8")  # explicit little-endian
         assert row.shape == (10,)
         assert row[0] == pytest.approx(1.5)     # timestamp
         assert row[1] == pytest.approx(0.01)    # position_x
@@ -180,7 +180,7 @@ class TestWriteHapticSample:
             )
         assert buf.tell() == 240  # 3 samples × 80 bytes
         buf.seek(0)
-        data = np.frombuffer(buf.read(), dtype=np.float64).reshape(-1, 10)
+        data = np.frombuffer(buf.read(), dtype="<f8").reshape(-1, 10)
         assert data.shape == (3, 10)
         for i in range(3):
             assert data[i, 0] == pytest.approx(float(i))  # timestamp
